+++ lib/src/picker/builder.dart
@@ -4,6 +4,7 @@ import 'package:flutter/material.dart';
 import 'package:provider/provider.dart';
 import 'package:vhcblade_theme/src/picker/bloc.dart';
 import 'package:vhcblade_theme/src/picker/event.dart';
+import 'package:vhcblade_theme/vhcblade_picker.dart';
 
 /// A builder that will automatically create a [VHCBladeThemeBloc] and provide it to the [builder] and the widget tree.
 ///
@@ -30,20 +31,32 @@ class _VHCBladeThemeBuilderState extends State<VHCBladeThemeBuilder> {
   @override
   Widget build(BuildContext context) {
     return BlocProvider<VHCBladeThemeBloc>(
-        create: (_, parentChannel) => VHCBladeThemeBloc(
+      create: (_, parentChannel) => VHCBladeThemeBloc(
+          databaseRepo: widget.databaseRepository ??
+              (widget.automaticallyRetrieveDatabaseRepository
+                  ? context.read<DatabaseRepository?>()
+                  : null),
+          parentChannel: parentChannel),
+      child: BlocProvider<VHCBladeUnlockThemeBloc>(
+        create: (read, parentChannel) => VHCBladeUnlockThemeBloc(
             databaseRepo: widget.databaseRepository ??
                 (widget.automaticallyRetrieveDatabaseRepository
                     ? context.read<DatabaseRepository?>()
                     : null),
+            bloc: read.read<VHCBladeThemeBloc>(),
             parentChannel: parentChannel),
-        child: Builder(builder: (context) {
-          if (!initialized) {
-            Future.delayed(Duration.zero).then((_) =>
-                context.fireEvent(VHCBladeThemeEvent.loadTheme.event, null));
-            initialized = true;
-          }
-          return widget.builder(context,
-              context.watchBloc<VHCBladeThemeBloc>().currentTheme.themeData);
-        }));
+        child: Builder(
+          builder: (context) {
+            if (!initialized) {
+              Future.delayed(Duration.zero).then((_) =>
+                  context.fireEvent(VHCBladeThemeEvent.loadTheme.event, null));
+              initialized = true;
+            }
+            return widget.builder(context,
+                context.watchBloc<VHCBladeThemeBloc>().currentTheme.themeData);
+          },
+        ),
+      ),
+    );
   }
 }